requirements.txt
bcrypt==4.1.2
cffi==1.16.0
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
fonttools==4.51.0
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
nengo==4.0.0
nengo-extras==0.5.0
nengo-fpga==0.2.3.dev0
nengo-gui==0.5.0
numpy==1.26.4
packaging==24.0
paramiko==3.4.0
pillow==10.3.0
pip==21.2.4
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==58.1.0
six==1.16.0
threadpoolctl==3.4.0
